# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: stocks
    description: Stock screener operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /stocks:
    get:
      operationId: getStocks
      tags: [stocks]
      summary: Get screener stocks
      description: Returns list of stocks with fundamental metrics for screening
      responses:
        "200":
          description: List of stocks with metrics
          content:
            application/json:
              schema:
                type: object
                properties:
                  stocks:
                    type: array
                    items:
                      $ref: "#/components/schemas/Stock"
                  cachedAt:
                    type: string
                    description: ISO timestamp of when the data was last fetched
                  source:
                    type: string
                    description: Data source identifier
                required:
                  - stocks
                  - cachedAt
                  - source
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "503":
          description: Service unavailable (API key not configured or quota exhausted)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Stock:
      type: object
      properties:
        ticker:
          type: string
        company:
          type: string
        sector:
          type: string
        marketCap:
          type: string
          enum: [Mega, Large, Mid, Small, Micro]
        epsGrowth5yr:
          type: number
          description: 5-year average EPS growth rate (decimal, e.g. 0.15 = 15%)
        consecutiveYearsAbove16:
          type: integer
          description: Number of consecutive recent years with EPS growth >= 16%
        pegRatio:
          type: number
        forwardPE:
          type: number
        revenueGrowth3yr:
          type: number
          description: 3-year average revenue growth rate (decimal)
        roe:
          type: number
          description: Return on equity (decimal)
        netMargin:
          type: number
          description: Net profit margin (decimal)
        debtToEquity:
          type: number
        yearsTo100x:
          type: number
        hundredBaggerScore:
          type: integer
      required:
        - ticker
        - company
        - sector
        - marketCap
        - epsGrowth5yr
        - consecutiveYearsAbove16
        - pegRatio
        - forwardPE
        - revenueGrowth3yr
        - roe
        - netMargin
        - debtToEquity
        - yearsTo100x
        - hundredBaggerScore
    ApiError:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message