# lib/api-zod/src/generated/types/stock.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StockMarketCap } from "./stockMarketCap";

export interface Stock {
  ticker: string;
  company: string;
  sector: string;
  marketCap: StockMarketCap;
  /** 5-year average EPS growth rate (decimal, e.g. 0.15 = 15%) */
  epsGrowth5yr: number;
  /** Number of consecutive recent years with EPS growth >= 16% */
  consecutiveYearsAbove16: number;
  pegRatio: number;
  forwardPE: number;
  /** 3-year average revenue growth rate (decimal) */
  revenueGrowth3yr: number;
  /** Return on equity (decimal) */
  roe: number;
  /** Net profit margin (decimal) */
  netMargin: number;
  debtToEquity: number;
  yearsTo100x: number;
  hundredBaggerScore: number;
}